Impelemted Reza's suggestion for classification on top of the BERT base model.  [Dense(n/2, 'linear')->BN()->Dense(n, 'linear')] is the same as covariance matrix decomposition. Relu would mess with that. It essentially estimates eigenvectors of the distribution of your items but selects the top n/2 eigenvectors for which the eigenvalues are the largest (the best separators). BN would play the role of removing the mean of the population (covariance is calculated on the centered data).
In general, you use Dense(?, linear)->BN->transfer function. The reason is that BN is supposed to redistribute your data around the center, and kind of in a normal distribution way with mean zero and variance 1. Now, if you apply a dense and that moves things away from the center and then apply your transfer function then you may end-up with an ill-posed distribution of your items, on which application of BN would not be effective. For example, assume your dense moves things to a far point from the center and you apply tanh as transfer. That would end up with all items mapped into all 1's for all dimensions. BN would map this to the center of the coordinate system with variance 0 (all items would be the same). This is ill-posed.